Reorganize Choices and related components
