+++ packages/lib-classifier/src/plugins/tasks/SurveyTask/components/components/Chooser/components/Choices/index.js
@@ -1,0 +1,1 @@
+export { default } from './Choices'
